Create catch-my-ear.md
I wrote this one some time in 2022 (its not dated in my book).  From a time when I recognised a deeper sense of nostalgia and longing arising in myself. I was wrestling with the grief of loss, days gone by, days that never came, and the passage of time. It reflects on the faint silence of my yesterdays, the echoes from the yesterdays that preceded mine and the continuous circling needle of the great song of life and death.